Auth fix

+++ WebApplication/src/main/java/com/cloud/csye6225/assignment/controller/BillController.java
@@ -10,10 +10,13 @@ import java.util.Collection;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -57,19 +60,27 @@ public class BillController {
 	@ResponseBody
 	public ResponseEntity<Bill> registerPost(@RequestBody String bill) {
 
-		Bill billObj = new Gson().fromJson(bill, Bill.class);
-		UserAccount account = accountService.currentUser;
+		if (SecurityContextHolder.getContext().getAuthentication() != null
+				&& SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken) {
+			// response.put("message", "you are not logged in!!!");
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 
-		billObj.setId(UUID.randomUUID().toString());
-		billObj.setCreated_ts(new Date().toString());
-		billObj.setUpdated_ts(new Date().toString());
-		billObj.setOwner_id(account.getId());
-		// billObj.setCategories(categories);
-		// billObj.setPaymentStatus(bill);
-		// billObj.setVendor(account.get);
 
-		billService.addBill(billObj);
-		return new ResponseEntity<>(billObj, HttpStatus.OK);
+		} else {
+			Bill billObj = new Gson().fromJson(bill, Bill.class);
+			UserAccount account = accountService.currentUser;
+
+			billObj.setId(UUID.randomUUID().toString());
+			billObj.setCreated_ts(new Date().toString());
+			billObj.setUpdated_ts(new Date().toString());
+			billObj.setOwner_id(account.getId());
+			// billObj.setCategories(categories);
+			// billObj.setPaymentStatus(bill);
+			// billObj.setVendor(account.get);
+
+			billService.addBill(billObj);
+			return new ResponseEntity<>(billObj, HttpStatus.OK);
+		}
 	}
 
 	// GEt all bills
@@ -79,9 +90,25 @@ public class BillController {
 //      if() return null;// log in or not
 		// get user id after log in
 
-		Collection<Bill> bills = billService.getAllBill();
+		Collection<Bill> bills1=null;
+		// check whether user logged in using the basic auth credentials or not
+		if (SecurityContextHolder.getContext().getAuthentication() != null
+				&& SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken) {
+			// response.put("message", "you are not logged in!!!");
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 
-		return new ResponseEntity<>(bills, HttpStatus.OK);
+		} else {
+
+			UserAccount account = accountService.currentUser;
+
+			Collection<Bill> bills = billService.getAllBill();
+
+			bills1 = bills.stream().filter(p -> p.getOwner_id().equals(account.getId()))
+					.collect(Collectors.toList());
+			return new ResponseEntity<>(bills1, HttpStatus.OK);
+
+		}
+		
 
 	}
 
@@ -92,21 +119,28 @@ public class BillController {
 //        if() return null;// log in or not
 		// get user id after log in
 
-		UserAccount account = accountService.currentUser;
+		if (SecurityContextHolder.getContext().getAuthentication() != null
+				&& SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken) {
+			// response.put("message", "you are not logged in!!!");
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 
-		if (billId == null) {
-			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-		}
+		} else {
 
-		else {
-			Bill bill = billService.getBillById(billId);
-			if (bill == null) {
-				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			UserAccount account = accountService.currentUser;
+
+			if (billId == null) {
+				return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 			}
 
-			return new ResponseEntity<>(bill, HttpStatus.OK);
-		}
+			else {
+				Bill bill = billService.getBillById(billId);
+				if (bill == null) {
+					return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+				}
 
+				return new ResponseEntity<>(bill, HttpStatus.OK);
+			}
+		}
 	}
 
 	// DELETE API
@@ -114,20 +148,32 @@ public class BillController {
 	@RequestMapping(value = "/v1/bill/{id}", method = RequestMethod.DELETE)
 	@ResponseBody
 	public ResponseEntity<?> deleteBillById(@PathVariable("id") String billId) {
-		boolean isSuccess = false;
-		if (billId == null) {
-			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-		} else {
-			Bill bill = billService.getBillById(billId);
-			if (bill == null) {
-				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-			}
-			isSuccess = billService.deleteBillById(billId);
 
-			if (isSuccess)
-				return new ResponseEntity<>("Deleted successfully", HttpStatus.OK);
-			else
-				return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			boolean isSuccess = false;
+			if (billId == null) {
+				return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+			} else {
+				
+				
+				Bill bill = billService.getBillById(billId);
+				if (SecurityContextHolder.getContext().getAuthentication() != null
+						&& SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken) {
+					// response.put("message", "you are not logged in!!!");
+					return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+
+				}
+				else{
+				
+				if (bill == null) {
+					return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+				}
+				isSuccess = billService.deleteBillById(billId);
+
+				if (isSuccess)
+					return new ResponseEntity<>("Deleted successfully", HttpStatus.OK);
+				else
+					return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			}
 		}
 	}
 
@@ -137,18 +183,24 @@ public class BillController {
 	public ResponseEntity<Bill> updateBill(@PathVariable("id") String billId, @RequestBody Bill infos) {
 //        if() return null;// log in or not
 		// get user id after log in
-
+		
 		Bill bill = billService.getBillById(billId);
+		if (SecurityContextHolder.getContext().getAuthentication() != null
+				&& SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken) {
+			// response.put("message", "you are not logged in!!!");
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 
-		bill.setOwner_id(infos.getOwner_id());
-		bill.setAmount_due(infos.getAmount_due());
-		bill.setVendor(infos.getVendor());
-		bill.setBill_date(infos.getBill_date());
-		bill.setDue_date(infos.getDue_date());
+		} else {
 
-		billService.updateBill(bill);
-		return new ResponseEntity<>(bill, HttpStatus.OK);
+			// bill.setOwner_id(infos.getOwner_id());
+			bill.setAmount_due(infos.getAmount_due());
+			bill.setVendor(infos.getVendor());
+			bill.setBill_date(infos.getBill_date());
+			bill.setDue_date(infos.getDue_date());
 
+			billService.updateBill(bill);
+			return new ResponseEntity<>(bill, HttpStatus.OK);
+		}
 	}
 
 }
